style: Demo파일 생성/불러오기, 확장자명 변경
