+++ src/components/lounge/LoungeDemoData.ts
@@ -1,0 +1,13 @@
+type TestDoorType = {
+  userNum: number;
+  userName: string;
+};
+
+const TestDoor: TestDoorType[] = [
+  { userNum: 1, userName: '수정' },
+  { userNum: 2, userName: '유라' },
+  { userNum: 3, userName: '소연' },
+  { userNum: 4, userName: '수진' },
+];
+
+export default TestDoor;
