+++ src/components/lounge/LoungeDoor.tsx
@@ -1,16 +1,6 @@
 import styled from 'styled-components';
 
-type TestDoorType = {
-  userNum: number;
-  userName: string;
-};
-
-const TestDoor: TestDoorType[] = [
-  { userNum: 1, userName: '수정' },
-  { userNum: 2, userName: '유라' },
-  { userNum: 3, userName: '소연' },
-  { userNum: 4, userName: '수진' },
-];
+import TestDoor from '~components/lounge/LoungeDemoData';
 
 const LoungeDoor = () => {
   return (
